requirements.txt
# Core Python packages
streamlit==1.35.0
requests

# Django and DRF
Django>=4.2
djangorestframework

# NLP & ML
transformers==4.40.1
torch
langchain==0.1.16
langchain-huggingface

# PDF/Text Parsing
PyMuPDF==1.24.1  # for PDF support

# Vector storage and embeddings
faiss-cpu
sentence-transformers
